requirements.txt
smbus2==0.4.1
bme280==0.3.3
mysql-connector-python==8.0.26
